Zoom in double

+++ src/combi_ckr_core.c
@@ -199,7 +199,7 @@ void compute_image(double *out, const double *buffer, int w, int h, int pd, int 
     }
 }
 
-static void fill_buffer(double *buffer, const double *image, double *homo, float zoom, int w, int h, int pd,
+static void fill_buffer(double *buffer, const double *image, double *homo, double zoom, int w, int h, int pd,
                         int order, int Nreg, double sigma2, int raw, const double *applic)
 {
     int wout = w*zoom;
@@ -336,7 +336,7 @@ static void fill_buffer(double *buffer, const double *image, double *homo, float
 }
 
 void fill_buffer_routine(const char *filename_image, char *filename_homo, int w, int h, int pd,
-                         double *buffer, float zoom, int order, int Nreg, double sigma2,
+                         double *buffer, double zoom, int order, int Nreg, double sigma2,
                          int raw, const double *applic)
 {
     /* read input image and verify the size */
@@ -360,7 +360,7 @@ void fill_buffer_routine(const char *filename_image, char *filename_homo, int w,
 void fill_buffer_routine_parallel(const char *filename_image, char *filename_homo,
                          const char *filename_image2, char *filename_homo2,
                          int w, int h, int pd,
-                         double *buffer, double *buffer2, float zoom, int order, int Nreg, double sigma2,
+                         double *buffer, double *buffer2, double zoom, int order, int Nreg, double sigma2,
                          int raw, const double *applic)
 {
     /* read input image and verify the size */
